+++ votor-quic/Cargo.toml
@@ -9,8 +9,13 @@ license = { workspace = true }
 edition = { workspace = true }
 publish = true
 
+[lib]
+
 [dependencies]
 anyhow = { workspace = true }
+arc-swap.workspace = true
+bytes.workspace = true
+crossbeam-channel.workspace = true
 futures-util = { workspace = true }
 pem.workspace = true
 quinn = { workspace = true }
